Removed RubyCritic and associated CLI option
While the promise of this gem was a good one -- and I still wish we had
better reporting for open source projects -- the gem has become a
burden. Especially since the gem is poorly maintained. By removing this
gem (and it's dependencies) an entire class of Ruby warnings will be
eliminated and the experience of working within this project will be
improved too.