Fixed README private gem server publising documentation
Some of this documentation was old and dated from earlier versions
where Gemsmith supported the `allowed_push_key` metadata key but that
is no longer supported. The documentation has been updated to better
reflect these changes and provide better usage examples.

Minor edits were made to the rest of the README to provide corrections
and clarifications.