Fixed RuboCop Metrics/CollectionLiteralLength
Necessary to improve code quality and reduce the lines of code required to read through. This has the added benefit of reducing duplication as well. I opted to mark `Gemsmith::CLI::Actions::Build` as a temporary TODO because I believe fixing how Rubysmith handles this first is the better solution.